Move hooks.py to _internal
