+++ certbot/_internal/plugins/manual.py
@@ -7,7 +7,7 @@ from acme.magic_typing import Dict  # pylint: disable=unused-import, no-name-in-
 
 from certbot import achallenges  # pylint: disable=unused-import
 from certbot import errors
-from certbot import hooks
+from certbot._internal import hooks
 from certbot import interfaces
 from certbot import reverter
 from certbot.compat import os
