+++ certbot/_internal/renewal.py
@@ -18,7 +18,7 @@ from acme.magic_typing import List  # pylint: disable=unused-import, no-name-in-
 from certbot import cli
 from certbot import crypto_util
 from certbot import errors
-from certbot import hooks
+from certbot._internal import hooks
 from certbot import interfaces
 from certbot._internal import storage
 from certbot._internal import updater
