--- docs/api/hooks.rst
@@ -1,5 +1,0 @@
-:mod:`certbot.hooks`
-------------------------
-
-.. automodule:: certbot.hooks
-   :members:
